Rename files and move Aggregate command
