+++ src/test/spec/unified_runner/operation/count.rs
@@ -1,98 +1,17 @@
 use crate::{
     error::Result,
-    options::{AggregateOptions, CountOptions, DistinctOptions, EstimatedDocumentCountOptions},
+    options::{CountOptions, DistinctOptions, EstimatedDocumentCountOptions},
     test::spec::unified_runner::{
         operation::{with_mut_session, with_opt_session, TestOperation},
         Entity,
         ExpectedEvent,
         TestRunner,
     },
-    Collection,
-    Database,
 };
 use bson::{Bson, Document};
-use futures::{future::BoxFuture, TryStreamExt};
+use futures::future::BoxFuture;
 use futures_util::FutureExt;
 use serde::Deserialize;
-
-#[derive(Debug, Deserialize)]
-pub(super) struct Aggregate {
-    pipeline: Vec<Document>,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: AggregateOptions,
-}
-
-impl TestOperation for Aggregate {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let result = match &self.session {
-                Some(session_id) => {
-                    enum AggregateEntity {
-                        Collection(Collection<Document>),
-                        Database(Database),
-                        Other(String),
-                    }
-                    let entity = match test_runner.entities.read().await.get(id).unwrap() {
-                        Entity::Collection(c) => AggregateEntity::Collection(c.clone()),
-                        Entity::Database(d) => AggregateEntity::Database(d.clone()),
-                        other => AggregateEntity::Other(format!("{:?}", other)),
-                    };
-                    with_mut_session!(test_runner, session_id, |session| async {
-                        let mut cursor = match entity {
-                            AggregateEntity::Collection(collection) => {
-                                collection
-                                    .aggregate(self.pipeline.clone())
-                                    .with_options(self.options.clone())
-                                    .session(&mut *session)
-                                    .await?
-                            }
-                            AggregateEntity::Database(db) => {
-                                db.aggregate(self.pipeline.clone())
-                                    .with_options(self.options.clone())
-                                    .session(&mut *session)
-                                    .await?
-                            }
-                            AggregateEntity::Other(debug) => {
-                                panic!("Cannot execute aggregate on {}", &debug)
-                            }
-                        };
-                        cursor.stream(session).try_collect::<Vec<Document>>().await
-                    })
-                    .await?
-                }
-                None => {
-                    let entities = test_runner.entities.read().await;
-                    let cursor = match entities.get(id).unwrap() {
-                        Entity::Collection(collection) => {
-                            collection
-                                .aggregate(self.pipeline.clone())
-                                .with_options(self.options.clone())
-                                .await?
-                        }
-                        Entity::Database(db) => {
-                            db.aggregate(self.pipeline.clone())
-                                .with_options(self.options.clone())
-                                .await?
-                        }
-                        other => panic!("Cannot execute aggregate on {:?}", &other),
-                    };
-                    cursor.try_collect::<Vec<Document>>().await?
-                }
-            };
-            Ok(Some(Bson::from(result).into()))
-        }
-        .boxed()
-    }
-
-    fn returns_root_documents(&self) -> bool {
-        true
-    }
-}
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
